Add eh error classes
