+++ sdk/eventhub/azure-eventhubs/azure/eventhub/aio/sender_async.py
@@ -12,7 +12,9 @@ from uamqp import SendClientAsync
 
 from azure.eventhub import MessageSendResult
 from azure.eventhub import EventHubError
-from azure.eventhub.common import _error_handler, EventData, _BatchSendEventData
+from azure.eventhub.common import EventData, _BatchSendEventData
+from azure.eventhub.error import EventHubError, EventHubConnectionError, \
+    EventHubAuthenticationError, EventHubMessageError, _error_handler
 
 log = logging.getLogger(__name__)
 
@@ -120,10 +122,36 @@ class Sender(object):
                 client_name=self.name,
                 properties=self.client.create_properties(self.client.config.user_agent),
                 loop=self.loop)
-        await self._handler.open_async()
-        self.running = True
-        while not await self._handler.client_ready_async():
-            await asyncio.sleep(0.05)
+        if not self.running:
+            try:
+                await self._handler.open_async()
+                self.running = True
+                while not await self._handler.client_ready_async():
+                    await asyncio.sleep(0.05)
+            except errors.AuthenticationException:
+                log.info("Sender failed authentication. Retrying...")
+                await self.reconnect()
+            except (errors.LinkDetach, errors.ConnectionClose) as shutdown:
+                if shutdown.action.retry and self.auto_reconnect:
+                    log.info("Sender detached. Attempting reconnect.")
+                    await self.reconnect()
+                else:
+                    log.info("Sender detached. Failed to connect")
+                    error = EventHubConnectionError(str(shutdown), shutdown)
+                    raise error
+            except errors.AMQPConnectionError as shutdown:
+                if str(shutdown).startswith("Unable to open authentication session") and self.auto_reconnect:
+                    log.info("Sender couldn't authenticate.", shutdown)
+                    error = EventHubAuthenticationError(str(shutdown))
+                    raise error
+                else:
+                    log.info("Sender connection error (%r).", shutdown)
+                    error = EventHubConnectionError(str(shutdown))
+                    raise error
+            except Exception as e:
+                log.info("Unexpected error occurred (%r)", e)
+                error = EventHubError("Sender connect failed: {}".format(e))
+                raise error
 
     async def _reconnect(self):
         await self._handler.close_async()
@@ -145,7 +173,7 @@ class Sender(object):
             return True
         except errors.TokenExpired as shutdown:
             log.info("AsyncSender disconnected due to token expiry. Shutting down.")
-            error = EventHubError(str(shutdown), shutdown)
+            error = EventHubAuthenticationError(str(shutdown), shutdown)
             await self.close(exception=error)
             raise error
         except (errors.LinkDetach, errors.ConnectionClose) as shutdown:
@@ -153,7 +181,7 @@ class Sender(object):
                 log.info("AsyncSender detached. Attempting reconnect.")
                 return False
             log.info("AsyncSender reconnect failed. Shutting down.")
-            error = EventHubError(str(shutdown), shutdown)
+            error = EventHubConnectionError(str(shutdown), shutdown)
             await self.close(exception=error)
             raise error
         except errors.MessageHandlerError as shutdown:
@@ -161,7 +189,7 @@ class Sender(object):
                 log.info("AsyncSender detached. Attempting reconnect.")
                 return False
             log.info("AsyncSender reconnect failed. Shutting down.")
-            error = EventHubError(str(shutdown), shutdown)
+            error = EventHubConnectionError(str(shutdown), shutdown)
             await self.close(exception=error)
             raise error
         except errors.AMQPConnectionError as shutdown:
@@ -169,7 +197,7 @@ class Sender(object):
                 log.info("AsyncSender couldn't authenticate. Attempting reconnect.")
                 return False
             log.info("AsyncSender connection error (%r). Shutting down.", shutdown)
-            error = EventHubError(str(shutdown))
+            error = EventHubConnectionError(str(shutdown))
             await self.close(exception=error)
             raise error
         except Exception as e:
@@ -219,14 +247,13 @@ class Sender(object):
         await self._handler.close_async()
 
     async def _send_event_data(self, event_data):
-        if not self.running:
-            await self._open()
+        await self._open()
         try:
             self._handler.send_message(event_data.message)
             if self._outcome != MessageSendResult.Ok:
                 raise Sender._error(self._outcome, self._condition)
         except errors.MessageException as failed:
-            error = EventHubError(str(failed), failed)
+            error = EventHubMessageError(str(failed), failed)
             await self.close(exception=error)
             raise error
         except (errors.TokenExpired, errors.AuthenticationException):
@@ -238,7 +265,7 @@ class Sender(object):
                 await self.reconnect()
             else:
                 log.info("Sender detached. Shutting down.")
-                error = EventHubError(str(shutdown), shutdown)
+                error = EventHubConnectionError(str(shutdown), shutdown)
                 await self.close(exception=error)
                 raise error
         except errors.MessageHandlerError as shutdown:
@@ -247,7 +274,7 @@ class Sender(object):
                 await self.reconnect()
             else:
                 log.info("Sender detached. Shutting down.")
-                error = EventHubError(str(shutdown), shutdown)
+                error = EventHubConnectionError(str(shutdown), shutdown)
                 await self.close(exception=error)
                 raise error
         except Exception as e:
@@ -266,10 +293,10 @@ class Sender(object):
             partition_key = ed.partition_key
             yield ed
         except StopIteration:
-            raise ValueError("batch_event_data must not be empty")
+            raise ValueError("event_data must not be empty")
         for ed in ed_iter:
             if ed.partition_key != partition_key:
-                raise ValueError("partition key of all EventData must be the same if being sent in a batch")
+                log.warning("partition key of all EventData must be the same if being sent in a batch")
             yield ed
 
     async def send(self, event_data):
@@ -301,70 +328,6 @@ class Sender(object):
             wrapper_event_data = _BatchSendEventData(self._verify_partition(event_data))
         wrapper_event_data.message.on_send_complete = self._on_outcome
         await self._send_event_data(wrapper_event_data)
-
-    def queue_message(self, event_data, callback=None):
-        """
-        Transfers an event data and notifies the callback when the operation is done.
-
-        :param event_data: The event to be sent.
-        :type event_data: ~azure.eventhub.common.EventData
-        :param callback: Callback to be run once the message has been send.
-         This must be a function that accepts two arguments.
-        :type callback: callable[~uamqp.constants.MessageSendResult, ~azure.eventhub.common.EventHubError]
-
-        Example:
-            .. literalinclude:: ../examples/test_examples_eventhub.py
-                :start-after: [START eventhub_client_transfer]
-                :end-before: [END eventhub_client_transfer]
-                :language: python
-                :dedent: 4
-                :caption: Transfers an event data and notifies the callback when the operation is done.
-
-        """
-        if self.error:
-            raise self.error
-        if not self.running:
-            self._open()
-        if event_data.partition_key and self.partition:
-            raise ValueError("EventData partition key cannot be used with a partition sender.")
-        if callback:
-            event_data.message.on_send_complete = lambda o, c: callback(o, Sender._error(o, c))
-        self._handler.queue_message(event_data.message)
-
-    async def send_pending_messages(self):
-        """
-        Wait until all transferred events have been sent.
-        """
-        if self.error:
-            raise self.error
-        if not self.running:
-            raise ValueError("Unable to send until client has been started.")
-        try:
-            await self._handler.wait_async()
-        except (errors.TokenExpired, errors.AuthenticationException):
-            log.info("AsyncSender disconnected due to token error. Attempting reconnect.")
-            await self.reconnect()
-        except (errors.LinkDetach, errors.ConnectionClose) as shutdown:
-            if shutdown.action.retry and self.auto_reconnect:
-                log.info("AsyncSender detached. Attempting reconnect.")
-                await self.reconnect()
-            else:
-                log.info("AsyncSender detached. Shutting down.")
-                error = EventHubError(str(shutdown), shutdown)
-                await self.close(exception=error)
-                raise error
-        except errors.MessageHandlerError as shutdown:
-            if self.auto_reconnect:
-                log.info("AsyncSender detached. Attempting reconnect.")
-                await self.reconnect()
-            else:
-                log.info("AsyncSender detached. Shutting down.")
-                error = EventHubError(str(shutdown), shutdown)
-                await self.close(exception=error)
-                raise error
-        except Exception as e:
-            log.info("Unexpected error occurred (%r).", e)
-            raise EventHubError("Send failed: {}".format(e))
 
     def _on_outcome(self, outcome, condition):
         """
